requirements.txt
celery==5.3.1
certifi==2023.7.22
chardet==5.2.0
Cython==3.0.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.54.0
django-configurations==2.4.1
django-crispy-forms==2.0
django-recaptcha==3.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
elasticsearch==8.9.0
expiringdict==1.2.2
gevent==23.7.0
greenlet==2.0.0
idna==3.4
ipython==8.14.0
kombu==5.3.1
mock==5.1.0
numpy==1.25.2
oauthlib==3.2.2
psycopg2-binary==2.9.7
pytest-cov==4.1.0
pytest-django==4.5.2
python-memcached==1.59
python-openid==2.2.5
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
word2vec==0.11.1